+++ .github/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2020-09-16
+## [v1.0.0] - 2020-09-16
 
 ### Added
 - Automatic download of Gecko and Chrome binaries upon install so that they don't slow down install time
@@ -21,4 +21,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Console command allowing users to start a local Selenium server.
 
+[v1.0.0]: https://github.com/Anteris-Dev/selenium-server/compare/v0.1.0...v1.0.0
 [v0.1.0]: https://github.com/Anteris-Dev/selenium-server/releases/tag/v0.1.0